Don't run raft network test with race detector
